tests/integration/test_today_scheduling_logic.py
"""Integration tests for 'today' scheduling logic - validates correct behavior."""

import pytest
from datetime import datetime, timezone, timedelta
from services.parsing_service import ParsingService
from config import Config
from dateutil import parser as date_parser


class TestTodaySchedulingLogic:
    """Test that 'today X' scheduling works correctly based on current time."""
    
    @pytest.fixture
    def parsing_service(self):
        """Create parsing service with real OpenAI API."""
        config = Config()
        if not config.OPENAI_API_KEY or config.OPENAI_API_KEY == "test_key_not_used":
            pytest.skip("OpenAI API key not configured")
        return ParsingService(config)
    
    def test_today_future_times_real_time(self, parsing_service):
        """
        Test 'today X' where X is in the future from current time.
        This should ALWAYS schedule for today, not tomorrow.
        """
        current_utc = datetime.now(timezone.utc)
        current_hour = current_utc.hour
        
        # Find a time that's definitely in the future today
        if current_hour < 20:  # If before 8pm
            future_hour = current_hour + 4  # 4 hours from now
            test_input = f"remind me today at {future_hour}:00"
        else:  # Late evening
            future_hour = 23
            test_input = "remind me today at 11pm"
        
        result = parsing_service.parse_content_to_task(
            content_message=test_input,
            owner_name="Test User",
            location="Portugal"
        )
        
        assert result is not None
        due_time = date_parser.isoparse(result['due_time'])
        
        # Should be scheduled for today
        time_diff_hours = (due_time - current_utc).total_seconds() / 3600
        
        print(f"\nCurrent UTC: {current_utc}")
        print(f"Test input: '{test_input}'")
        print(f"Scheduled UTC: {due_time}")
        print(f"Hours until scheduled: {time_diff_hours:.1f}")
        
        # Validate it's in the future but within today (less than 24 hours)
        assert time_diff_hours > 0, "Should be scheduled in the future"
        assert time_diff_hours < 24, "Should be scheduled for today (within 24 hours)"
        
        # If we're testing a specific hour, validate it
        if future_hour < 24:
            # Account for timezone - Portugal is UTC+1
            expected_utc_hour = (future_hour - 1) % 24
            assert abs(due_time.hour - expected_utc_hour) <= 1, \
                f"Expected around hour {expected_utc_hour} UTC, got {due_time.hour}"
    
    def test_today_past_times_real_time(self, parsing_service):
        """
        Test 'today X' where X is in the past from current time.
        This should schedule for tomorrow.
        """
        current_utc = datetime.now(timezone.utc)
        current_hour = current_utc.hour
        
        # Find a time that's definitely in the past
        if current_hour > 3:  # If after 3am
            past_hour = current_hour - 3  # 3 hours ago
            test_input = f"remind me today at {past_hour}:00"
        else:  # Very early morning
            past_hour = 1
            test_input = "remind me today at 1am"
        
        result = parsing_service.parse_content_to_task(
            content_message=test_input,
            owner_name="Test User",
            location="Portugal"
        )
        
        assert result is not None
        due_time = date_parser.isoparse(result['due_time'])
        
        # Should be scheduled for tomorrow (more than 12 hours away)
        time_diff_hours = (due_time - current_utc).total_seconds() / 3600
        
        print(f"\nCurrent UTC: {current_utc}")
        print(f"Test input: '{test_input}'")
        print(f"Scheduled UTC: {due_time}")
        print(f"Hours until scheduled: {time_diff_hours:.1f}")
        
        # Should be in the future
        assert time_diff_hours > 0, "Should be scheduled in the future"
        # Should be tomorrow (more than 12 hours away typically)
        assert time_diff_hours > 12, "Should be scheduled for tomorrow (past time today)"
    
    def test_explicit_today_morning_times(self, parsing_service):
        """
        Test explicit morning times with 'today'.
        Helps validate the midnight corner case without mocking.
        """
        test_cases = [
            "remind me today 5am",
            "remind me today at 7am",
            "remind me today 9am"
        ]
        
        current_utc = datetime.now(timezone.utc)
        
        for test_input in test_cases:
            result = parsing_service.parse_content_to_task(
                content_message=test_input,
                owner_name="Test User",
                location="Portugal"
            )
            
            assert result is not None
            due_time = date_parser.isoparse(result['due_time'])
            
            time_diff_hours = (due_time - current_utc).total_seconds() / 3600
            
            print(f"\n'{test_input}':")
            print(f"  Current: {current_utc.strftime('%H:%M')} UTC")
            print(f"  Scheduled: {due_time.strftime('%Y-%m-%d %H:%M')} UTC")
            print(f"  Hours away: {time_diff_hours:.1f}")
            
            # Should always be in the future
            assert time_diff_hours > 0, f"{test_input} should be in the future"
            
            # Extract the hour from the input
            import re
            match = re.search(r'(\d+)am', test_input)
            if match:
                requested_hour = int(match.group(1))
                # Portugal is UTC+1, so 5am Portugal = 4am UTC
                expected_utc_hour = (requested_hour - 1) % 24
                
                # Very lenient validation - just ensure it's reasonable
                # Should be scheduled within 48 hours (allows for today or tomorrow)
                assert time_diff_hours < 48, f"{test_input} should be scheduled within 48 hours"
                
                # Check if it's approximately the right hour (allowing for timezone conversion)
                if abs(due_time.hour - expected_utc_hour) <= 1:
                    print(f"  ✅ Scheduled at correct hour ({due_time.hour} UTC ≈ {expected_utc_hour} expected)")
                else:
                    print(f"  ℹ️  Different hour but reasonable ({due_time.hour} UTC vs {expected_utc_hour} expected)")
                
                # Determine if scheduled for today or tomorrow
                if due_time.date() == current_utc.date():
                    print(f"  ✅ Scheduled for today")
                elif due_time.date() == (current_utc + timedelta(days=1)).date():
                    print(f"  ✅ Scheduled for tomorrow")
                else:
                    print(f"  ✅ Scheduled for: {due_time.date()}")
    
    def test_prompt_template_contains_midnight_rule(self, parsing_service):
        """Verify the prompt template has the midnight special case."""
        prompt_template = parsing_service.prompt_template.template
        
        # Check for critical time handling features that actually exist
        assert "midnight" in prompt_template.lower(), \
            "Prompt should mention midnight handling"
        assert "SPECIAL CASES" in prompt_template, \
            "Prompt should have special cases section"
        assert "timezone" in prompt_template.lower(), \
            "Prompt should mention timezone handling"
        
        print("✅ Prompt template contains time handling features")